Move wait events into a separate file
